+++ src/ts/events/hacks/Hack.ts
@@ -100,7 +100,7 @@ abstract class Hack {
     public static create(type?: number, level?: number): void {
         // Get a random hack type
         if (type === undefined) {
-            type = Utils.random(0, 6);
+            type = Utils.random(0, 7);
         }
 
         // Get the hack level
@@ -114,19 +114,22 @@ abstract class Hack {
                 new Cryptogram(level);
                 break;
             case 1:
-                new HexMatcher(level);
+                new DataCorruption(level);
                 break;
             case 2:
+                new HexMatcher(level);
+                break;
+            case 3:
                 if (Settings.isSettingEnabled("poor-eyesight-features")) {
                     Hack.create();
                 } else {
                     new HiddenPasswords(level);
                 }
                 break;
-            case 3:
+            case 4:
                 new LogMismatch(level);
                 break;
-            case 4:
+            case 5:
                 new NumberMultiples(level);
                 break;
             default: